+++ app/controllers/links_controller.rb
@@ -4,6 +4,7 @@ class LinksController < ApplicationController
   before_action :authenticate_user!
 
   def destroy
+    authorize link
     link.destroy if current_user.author_of?(link.linkable)
   end
